+++ authomatic/providers/oauth1.py
@@ -809,6 +809,7 @@ class Twitter(OAuth1):
 
     Supported :class:`.User` properties:
 
+    * email
     * city
     * country
     * id
@@ -822,7 +823,6 @@ class Twitter(OAuth1):
     Unsupported :class:`.User` properties:
 
     * birth_date
-    * email
     * gender
     * first_name
     * last_name
@@ -837,6 +837,7 @@ class Twitter(OAuth1):
         city=True,
         country=True,
         id=True,
+        email=True,
         link=True,
         locale=True,
         location=True,
@@ -848,7 +849,7 @@ class Twitter(OAuth1):
     request_token_url = 'https://api.twitter.com/oauth/request_token'
     user_authorization_url = 'https://api.twitter.com/oauth/authenticate'
     access_token_url = 'https://api.twitter.com/oauth/access_token'
-    user_info_url = 'https://api.twitter.com/1.1/account/verify_credentials.json?include_entities=true'
+    user_info_url = 'https://api.twitter.com/1.1/account/verify_credentials.json?include_entities=true&include_email=true'
     
     supports_jsonp = True